+++ @xen-orchestra/lite/src/pages/xoa-deploy.vue
@@ -191,7 +191,6 @@
 import FormInput from '@/components/form/FormInput.vue'
 import FormSection from '@/components/form/FormSection.vue'
 import TitleBar from '@/components/TitleBar.vue'
-import UiCard from '@/components/ui/UiCard.vue'
 import UiRaw from '@/components/ui/UiRaw.vue'
 import { usePageTitleStore } from '@/stores/page-title.store'
 import { useNetworkStore } from '@/stores/xen-api/network.store'
@@ -202,6 +201,7 @@ import VtsIcon from '@core/components/icon/VtsIcon.vue'
 import VtsInputWrapper from '@core/components/input-wrapper/VtsInputWrapper.vue'
 import VtsSelect from '@core/components/select/VtsSelect.vue'
 import UiButton from '@core/components/ui/button/UiButton.vue'
+import UiCard from '@core/components/ui/card/UiCard.vue'
 import UiRadioButton from '@core/components/ui/radio-button/UiRadioButton.vue'
 import UiRadioButtonGroup from '@core/components/ui/radio-button-group/UiRadioButtonGroup.vue'
 import UiToggle from '@core/components/ui/toggle/UiToggle.vue'
@@ -209,9 +209,9 @@ import { useFormSelect } from '@core/packages/form-select'
 import { useModal } from '@core/packages/modal/use-modal.ts'
 import { useUiStore } from '@core/stores/ui.store'
 import { logicNot } from '@vueuse/math'
-import { computed, ref } from 'vue'
+import { computed, onUnmounted, ref, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
-import { useRouter } from 'vue-router'
+import { onBeforeRouteLeave, useRouter } from 'vue-router'
 
 const REQUIRED_GB = 20
 
@@ -253,6 +253,41 @@ const status = ref<string | undefined>()
 const error = ref<string | undefined>()
 const url = ref<string | undefined>()
 const vmRef = ref<string | undefined>()
+
+const beforeWindowUnload = (event: BeforeUnloadEvent) => {
+  event.preventDefault()
+}
+
+onBeforeRouteLeave(async () => {
+  if (deploying.value) {
+    if (!window.confirm(t('do-you-really-want-to-leave-this-page?'))) {
+      return false
+    }
+
+    window.removeEventListener('beforeunload', beforeWindowUnload)
+    cancel()
+    return true
+  }
+})
+
+watch(
+  deploying,
+  deploying => {
+    if (deploying) {
+      window.addEventListener('beforeunload', beforeWindowUnload)
+    } else {
+      window.removeEventListener('beforeunload', beforeWindowUnload)
+    }
+  },
+  { flush: 'sync' }
+)
+
+onUnmounted(() => {
+  window.removeEventListener('beforeunload', beforeWindowUnload)
+  if (deploying.value) {
+    cancel()
+  }
+})
 
 const resetValues = () => {
   deploying.value = false
@@ -474,6 +509,8 @@ async function deploy() {
   } catch (err: any) {
     console.error(err)
     error.value = err?.message ?? err?.code ?? 'Unknown error'
+  } finally {
+    deploying.value = false
   }
 }