+++ @xen-orchestra/web-core/lib/tables/column-sets/traffic-rules-columns.ts
@@ -1,7 +1,7 @@
 import { defineColumns } from '@core/packages/table/define-columns.ts'
+import { useActionColumn } from '@core/tables/column-definitions/action-column.ts'
 import { useDoubleLinkColumn } from '@core/tables/column-definitions/double-link-column.ts'
 import { useNumberColumn } from '@core/tables/column-definitions/number-column.ts'
-import { useSelectItemColumn } from '@core/tables/column-definitions/select-item-column.ts'
 import { useTagColumn } from '@core/tables/column-definitions/tag-column.ts'
 import { useTextColumn } from '@core/tables/column-definitions/text-column.ts'
 import { useTruncatedTextColumn } from '@core/tables/column-definitions/truncated-text-column.ts'
@@ -11,19 +11,17 @@ export const useTrafficRulesColumns = defineColumns(() => {
   const { t } = useI18n()
 
   return {
-    order: useNumberColumn({ headerIcon: 'fa:hashtag' }),
+    order: useNumberColumn(),
     policy: useTagColumn({ headerLabel: () => t('policy') }),
     protocol: useTextColumn({
       headerLabel: () => t('protocol:port'),
-      headerIcon: 'fa:square-caret-down',
     }),
-    directionA: useTextColumn({ headerLabel: () => t('direction'), headerIcon: 'fa:square-caret-down' }),
-    target: useTruncatedTextColumn({ headerLabel: () => t('target'), headerIcon: 'fa:at' }),
+    directionA: useTextColumn({ headerLabel: () => t('direction') }),
+    target: useTruncatedTextColumn({ headerLabel: () => t('target') }),
     directionB: useTextColumn({
       headerLabel: () => t('direction'),
-      headerIcon: 'fa:square-caret-down',
     }),
     object: useDoubleLinkColumn({ headerLabel: () => t('object') }),
-    selectItem: useSelectItemColumn(),
+    actions: useActionColumn({}),
   }
 })