+++ @xen-orchestra/web/src/modules/storage-repository/components/list/panel/StorageRepositorySidePanel.vue
@@ -1,39 +1,45 @@
 <template>
-  <VtsStateHero v-if="!isReady" format="panel" type="busy" size="medium" />
+  <VtsStateHero v-if="!isReady" format="panel" size="medium" type="busy" />
   <UiPanel v-else :key="panelSignature" :class="{ 'mobile-drawer': uiStore.isSmall }">
     <template #header>
-      <VtsDeleteButton
-        class="sr-delete-button"
-        :disabled="!canDeleteSr"
-        :busy="isDeletingSr"
-        @click="openSrDeleteModal()"
-      />
-      <SrDisconnectButton class="sr-disconnect-button" :sr :scope />
+      <div class="action-buttons">
+        <SrConnectButton :scope :sr />
+        <MenuList placement="bottom-end">
+          <template #trigger="{ open }">
+            <UiButtonIcon accent="brand" icon="action:more-actions" size="medium" @click="open($event)" />
+          </template>
+          <SrDisconnectButton :scope :sr />
+          <SrDeleteButton :sr />
+        </MenuList>
+      </div>
+
       <div :class="{ 'action-buttons-container': uiStore.isSmall }">
         <UiButtonIcon
           v-tooltip="t('action:close')"
+          :icon="uiStore.isSmall ? 'fa:angle-left' : 'fa:close'"
+          accent="brand"
           size="small"
           variant="tertiary"
-          accent="brand"
-          :icon="uiStore.isSmall ? 'fa:angle-left' : 'fa:close'"
           @click="emit('close')"
         />
       </div>
     </template>
     <template #default>
-      <StorageRepositoryInfosCard :sr :scope />
+      <StorageRepositoryInfosCard :scope :sr />
       <StorageRepositorySpaceCard :sr />
-      <StorageRepositoryVdisCard :vdis />
+      <StorageRepositoryVdisCard :vdi-snapshots :vdis />
       <StorageRepositoryHostsCard :hosts />
-      <StorageRepositoryPbdsCard :sr :scope />
+      <StorageRepositoryPbdsCard :scope :sr />
       <StorageRepositoryCustomFieldsCard :custom-fields />
     </template>
   </UiPanel>
 </template>
 
-<script setup lang="ts">
-import { useXoHostCollection, type FrontXoHost } from '@/modules/host/remote-resources/use-xo-host-collection.ts'
+<script lang="ts" setup>
+import { type FrontXoHost, useXoHostCollection } from '@/modules/host/remote-resources/use-xo-host-collection.ts'
 import { useXoPbdCollection } from '@/modules/pbd/remote-resources/use-xo-pbd-collection.ts'
+import SrConnectButton from '@/modules/storage-repository/components/actions/connect/SrConnectButton.vue'
+import SrDeleteButton from '@/modules/storage-repository/components/actions/delete/SrDeleteButton.vue'
 import SrDisconnectButton from '@/modules/storage-repository/components/actions/disconnect/SrDisconnectButton.vue'
 import StorageRepositoryCustomFieldsCard from '@/modules/storage-repository/components/list/panel/cards/StorageRepositoryCustomFieldsCard.vue'
 import StorageRepositoryHostsCard from '@/modules/storage-repository/components/list/panel/cards/StorageRepositoryHostsCard.vue'
@@ -41,12 +47,15 @@ import StorageRepositoryInfosCard from '@/modules/storage-repository/components/
 import StorageRepositoryPbdsCard from '@/modules/storage-repository/components/list/panel/cards/StorageRepositoryPbdsCard.vue'
 import StorageRepositorySpaceCard from '@/modules/storage-repository/components/list/panel/cards/StorageRepositorySpaceCard.vue'
 import StorageRepositoryVdisCard from '@/modules/storage-repository/components/list/panel/cards/StorageRepositoryVdisCard.vue'
-import { useSrDeleteModal } from '@/modules/storage-repository/composables/use-sr-delete-modal.composable.ts'
 import { useGetPbdsInScope } from '@/modules/storage-repository/composables/xo-sr-utils.composable.ts'
 import type { FrontXoSr } from '@/modules/storage-repository/remote-resources/use-xo-sr-collection.ts'
-import type { StorageScope } from '@/modules/storage-repository/types/storage-scope.type.ts'
+import type { SrScope } from '@/modules/storage-repository/types/storage-repository.type'
 import { useXoVdiCollection } from '@/modules/vdi/remote-resources/use-xo-vdi-collection.ts'
-import VtsDeleteButton from '@core/components/delete-button/VtsDeleteButton.vue'
+import {
+  type FrontXoVdiSnapshot,
+  useXoVdiSnapshotCollection,
+} from '@/modules/vdi/remote-resources/use-xo-vdi-snapshot-collection.ts'
+import MenuList from '@core/components/menu/MenuList.vue'
 import VtsStateHero from '@core/components/state-hero/VtsStateHero.vue'
 import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
 import UiPanel from '@core/components/ui/panel/UiPanel.vue'
@@ -59,7 +68,7 @@ import { useI18n } from 'vue-i18n'
 
 const { sr, scope } = defineProps<{
   sr: FrontXoSr
-  scope: StorageScope
+  scope: SrScope
 }>()
 
 const emit = defineEmits<{
@@ -72,10 +81,13 @@ const uiStore = useUiStore()
 const { useGetVdisByIds, areVdisReady } = useXoVdiCollection()
 const { getHostById, areHostsReady } = useXoHostCollection()
 const { pbdsBySr, arePbdsReady } = useXoPbdCollection()
+const { useGetVdiSnapshotsByIds, areVdiSnapshotsReady } = useXoVdiSnapshotCollection()
 
-const isReady = logicAnd(areVdisReady, areHostsReady, arePbdsReady)
+const isReady = logicAnd(areVdisReady, areHostsReady, arePbdsReady, areVdiSnapshotsReady)
 
 const vdis = useGetVdisByIds(() => sr.VDIs as XoVdi['id'][])
+
+const vdiSnapshots = useGetVdiSnapshotsByIds(() => sr.VDIs as FrontXoVdiSnapshot['id'][])
 
 const { getSrPbdsSignature } = useGetPbdsInScope()
 
@@ -106,15 +118,14 @@ const customFields = computed(() => {
     return acc
   }, {})
 })
-
-const { openModal: openSrDeleteModal, canRun: canDeleteSr, isRunning: isDeletingSr } = useSrDeleteModal(() => [sr])
 </script>
 
-<style scoped lang="postcss">
-.sr-disconnect-button {
+<style lang="postcss" scoped>
+.action-buttons {
+  display: flex;
+  align-items: center;
   margin-inline-end: auto;
 }
-
 .mobile-drawer {
   position: fixed;
   inset: 0;